int -> long long int

+++ cppfuncs/index.cpp
@@ -5,26 +5,26 @@
 #endif
 
 namespace index {
-    int index2(int i1,int i2,int N1,int N2){
+    long long int index2(long long int i1,long long int i2,long long int N1,long long int N2){
         return i2 + i1*N2;
     }
-    int index3(int i1,int i2,int i3,int N1,int N2, int N3){
+    long long int index3(long long int i1,long long int i2,long long int i3,long long int N1,long long int N2, long long int N3){
         return i3 + i2*N3 + i1*N2*N3;
     }
-    int index4(int i1,int i2,int i3,int i4,int N1,int N2, int N3, int N4){
+    long long int index4(long long int i1,long long int i2,long long int i3,long long int i4,long long int N1,long long int N2, long long int N3, long long int N4){
         return i4 + (i3 + (i2 + i1*N2)*N3)*N4;
     }
 
-    int couple(int t,int iP,int iL,int iA,par_struct* par){
+    long long int couple(long long int t,long long int iP,long long int iL,long long int iA,par_struct* par){
         return index4(t,iP,iL,iA , par->T,par->num_power,par->num_love,par->num_A); 
     }
-    int couple_pd(int t,int iP,int iL,int iA_pd,par_struct* par){
+    long long int couple_pd(long long int t,long long int iP,long long int iL,long long int iA_pd,par_struct* par){
         return index4(t,iP,iL,iA_pd , par->T,par->num_power,par->num_love,par->num_A_pd); 
     }
-    int single_to_couple(int t,int iL,int iA,par_struct* par){
+    long long int single_to_couple(long long int t,long long int iL,long long int iA,par_struct* par){
         return index3(t,iL,iA , par->T,par->num_love,par->num_A); 
     }
-    int single(int t,int iA,par_struct* par){
+    long long int single(long long int t,long long int iA,par_struct* par){
         return index2(t,iA , par->T,par->num_A); 
     }
 
@@ -33,7 +33,7 @@ namespace index {
             int iL;
             int iA;
             par_struct *par; 
-            int idx(int iP){
+            long long int idx(long long int iP){
                     return index::couple(t,iP,iL,iA , par); 
             }
